requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
streamlit>=1.28.0
pandas>=2.0.0
PyPDF2>=3.0.0
python-docx>=0.8.10
requests>=2.25.1
python-dotenv>=0.19.0
qdrant-client>=1.6.0
sentence-transformers>=2.2.0
google-generativeai>=0.3.0
python-multipart>=0.0.5
docx2txt>=0.8
openpyxl>=3.0.0


//change model//
//metadata, system prompt and give functionality to the model//
